# src/main/java/com/lexel/elasticsearchdemojava/basicapp/model/Product.java
package com.lexel.elasticsearchdemojava.basicapp.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Product {

    private String id;
    private String name;
    private Double price;

}