# src/main/java/com/lexel/elasticsearchdemojava/basicapp/model/ProductResponse.java
package com.lexel.elasticsearchdemojava.basicapp.model;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@JsonInclude(JsonInclude.Include.NON_NULL)
public class ProductResponse {
    private String id;
    private String name;
    private Double price;
    private String message;
}